Update homedac_box.py
Added some comments on a bug + places where we need to check for errors. It will be important that we maintain 6 digit/20 bit precision so please check the strings you're sending to the dac box and make sure we aren't truncating/rounding the numbers. 

We will also want to have some dictionary mapping channel numbers to names so that we can specify for example dac.sweep2d("p1", "b1", 1.025, 1.030, 50, 0.625, 0.660, 50). We will also need to save somewhere (probably in same dictionary) parameters for each channel. i.e. each channel should have a voltage multiplier (we will have voltage dividers in the filter boxes that we want to be able to account for), There should also be software limits for maximum and minimum voltage applies (i.e. you shouldn't be able to apply more than 0.8V on any given gate after taking into account multipliers, etc.).

We still need to add in data collection for every step in the loop.

So, we will need to
- fix bug when generating 2d sweeps
- check precision of voltage sent.
- add in dictionary of parameters
- add function in sweeper that checks input strings against dictionary entries and returns channel value for programming dac